chore: added update keys for k8s events
